% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-anova.R
\name{summarize_column}
\alias{summarize_column}
\title{function name}
\usage{
summarize_column(data, col)
}
\arguments{
\item{data}{dataframe}

\item{col}{col name}
}
\value{
list
}
\description{
function description
}
\examples{
summarize_column(mtcars, "mpg")
}
